requirements.txt
requests==2.24.0
google-auth==1.23.0
pyyaml==5.3.1
